[server] | refactor(): 코드 리팩토링
각 모듈에있는 constants를 common의 constants로 옮겼습니다.
pipe에서 page validation에 필요한 정규식을 regexes.ts에 작성하였습니다.
