+++ packages/server/src/common/regexes.ts
@@ -1,0 +1,1 @@
+export const NATURAL_NUMBER = /^[1-9]\d*$/;
