--- packages/server/src/tag/constants.ts
@@ -1,7 +1,0 @@
-export const TAG_ITEMS_PER_PAGE = 24;
-export const SEARCHED_ITEM_NUMBER = 5;
-export const TAG_QUERY_SELECT_COLUMNS = [
-  'Tag.id',
-  'Tag.name',
-  'Tag.videosCount',
-];
