--- packages/server/src/tag/dto/tag-request-query.dto.ts
@@ -1,6 +1,0 @@
-export class TagListRequestQueryDto {
-  public constructor(
-    public readonly page: string,
-    public readonly keyword?: string,
-  ) {}
-}
