--- packages/server/src/user/constants.ts
@@ -1,7 +1,0 @@
-export const USER_ITEMS_PER_PAGE = 20;
-export const SEARCHED_ITEM_NUMBER = 5;
-export const USER_QUERY_SELECT_COLUMNS = [
-  'User.id',
-  'User.username',
-  'User.avatar',
-];
